Build the Edit Issue Page
Refactor: Build IssueForm component for NewIssue & EditIssue
Add loading skeleton
