+++ app/issues/[id]/edit/page.tsx
@@ -1,9 +1,20 @@
-import React from 'react'
+import React from "react";
+import IssueForm from "@/app/issues/components/IssueForm";
+import prisma from "@/prisma/client";
+import { notFound } from "next/navigation";
 
-const EditIssuePage = () => {
-  return (
-    <div>EditIssuePage</div>
-  )
+interface Props {
+  params: { id: string };
 }
 
-export default EditIssuePage
+const EditIssuePage = async ({ params }: Props) => {
+  const editIssue = await prisma.issue.findUnique({
+    where: { id: parseInt(params.id) },
+  });
+
+  if (!editIssue) notFound();
+
+  return <IssueForm issue={editIssue} />;
+};
+
+export default EditIssuePage;
