+++ app/issues/components/IssueForm.tsx
@@ -1,0 +1,78 @@
+"use client";
+
+import ErrorMessage from "@/app/components/ErrorMessage";
+import { createIssueSchema } from "@/app/validationSchema";
+import { zodResolver } from "@hookform/resolvers/zod";
+import { Issue } from "@prisma/client";
+import { Button, Callout, Spinner, TextField } from "@radix-ui/themes";
+import axios from "axios";
+import "easymde/dist/easymde.min.css";
+import dynamic from "next/dynamic";
+import { useRouter } from "next/navigation";
+import { useState } from "react";
+import { Controller, useForm } from "react-hook-form";
+import { z } from "zod";
+
+const SimpleMdeReact = dynamic(() => import("react-simplemde-editor"), {
+  ssr: false,
+});
+
+type IssueFormData = z.infer<typeof createIssueSchema>;
+
+const IssueForm = ({ issue }: { issue?: Issue }) => {
+  const { push } = useRouter();
+  const {
+    register,
+    control,
+    handleSubmit,
+    formState: { errors },
+  } = useForm<IssueFormData>({
+    resolver: zodResolver(createIssueSchema),
+  });
+  const [error, setError] = useState("");
+  const [loadingSubmit, setLoadingSubmit] = useState(false);
+
+  const createSubmitIssue = handleSubmit(async (data: IssueFormData) => {
+    try {
+      setLoadingSubmit(true);
+      await axios.post("/api/issues", data);
+      push("/issues");
+    } catch {
+      setLoadingSubmit(false);
+      setError("An unexpected error occurred. Please try again.");
+    }
+  });
+
+  return (
+    <div className="max-w-xl">
+      {error && (
+        <Callout.Root color="red" className="mb-5">
+          <Callout.Text>{error}</Callout.Text>
+        </Callout.Root>
+      )}
+      <form className="space-y-3" onSubmit={createSubmitIssue}>
+        <TextField.Root
+          defaultValue={issue?.title}
+          placeholder="Title"
+          {...register("title")}
+        />
+        <ErrorMessage>{errors.title?.message}</ErrorMessage>
+        <Controller
+          name="description"
+          control={control}
+          defaultValue={issue?.description}
+          render={({ field }) => (
+            <SimpleMdeReact placeholder="Description" {...field} />
+          )}
+        />
+        <ErrorMessage>{errors.description?.message}</ErrorMessage>
+        <Button disabled={loadingSubmit}>
+          Submit New Issue
+          <Spinner loading={loadingSubmit} />
+        </Button>
+      </form>
+    </div>
+  );
+};
+
+export default IssueForm;
